break up square into smaller files
